Use some more standard types and remove the need to use gtypes.h directly.
The header file gtypes.h is still needed indirectly because some of the
custom types like GooString and similar include that.

Substituting those custom types with some of the ones from standard
libraries (e.g., std::string) is a probably good idea, since they are
likely to be better implemented (perhaps with assembly versions and with
an eye towards the diversity of platforms).