Remove some unneeded checks before deallocating memory.
The standard guarantees that it is safe to deallocate memory with the
delete operator even if we have a `NULL` pointer. Therefore, we make the
code smaller without those checks.

On top of that, properly separate (with white space) some casts from
their "arguments".

There are many other interesting things that can be found in this code
with the help of [cppcheck][0].

[0]: http://cppcheck.sf.net/